backend-python/requirements.txt
black==24.4.2
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.2
cryptography==43.0.3
cycler==0.12.1
et_xmlfile==2.0.0
Flask==2.3.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.1.1
fonttools==4.58.4
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.8.4
mypy-extensions==1.0.0
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
redis==3.5.3
requests==2.32.3
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
xlrd==2.0.1
APScheduler==3.11.0
importlib-metadata==8.7.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5
